requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMuPDF==1.25.5
python-dotenv==1.1.0
Werkzeug==3.1.3
gunicorn==20.1.0
python-docx==1.1.0
